TT-2012: Provision the MSA EC2 instance
Note that we're committing some private keys here - this is deliberate -
the keys will not be used for anything other than testing.

The provisioning files we copy in are pretty much the same as the ones
in `/vagrant` - later we might want to think about sharing these between
terraform and vagrant.

We switched from using the ec2 module to using and aws_instance resource
because we couldn't work out how to add file provisioners to the module.

Co-authored-by: Rob Larsen <rob.larsen@digital.cabinet-office.gov.uk>